100% coverage on setup

+++ backend/internal/models/cart.model.go
@@ -4,6 +4,6 @@ import "go.mongodb.org/mongo-driver/bson/primitive"
 
 type Cart struct {
 	ID       primitive.ObjectID `bson:"_id" json:"id"`
-	Total    int64              `bson:"total" json:"total"`
+	Total    float64            `bson:"total" json:"total"`
 	Articles []Article          `bson:"articles" json:"articles"`
 }
